# Conflicts:
#	src/main/java/com/example/retro_care/customer/controller/CustomerController.java
#	src/main/java/com/example/retro_care/customer/dto/CustomerDto.java
#	src/main/java/com/example/retro_care/customer/dto/ICustomerDto.java
#	src/main/java/com/example/retro_care/customer/repository/ICustomerRepository.java
#	src/main/java/com/example/retro_care/customer/service/CustomerService.java
#	src/main/java/com/example/retro_care/customer/service/ICustomerService.java
#	src/main/java/com/example/retro_care/employee/controller/EmployeeController.java
#	src/main/java/com/example/retro_care/employee/repository/IEmployeeRepository.java
#	src/main/java/com/example/retro_care/employee/service/EmployeeService.java
#	src/main/java/com/example/retro_care/employee/service/IEmployeeService.java
#	src/main/java/com/example/retro_care/indication/controller/IndicationController.java
#	src/main/java/com/example/retro_care/indication/repository/IIndicationRepository.java
#	src/main/java/com/example/retro_care/indication/service/IIndicationService.java
#	src/main/java/com/example/retro_care/indication/service/IndicationService.java
#	src/main/java/com/example/retro_care/invoice/controller/InvoiceController.java
#	src/main/java/com/example/retro_care/invoice/model/InvoiceDetail.java
#	src/main/java/com/example/retro_care/invoice/model/InvoiceDetailDto.java
#	src/main/java/com/example/retro_care/invoice/model/InvoiceDto.java
#	src/main/java/com/example/retro_care/invoice/repository/IInvoiceDetailRepository.java
#	src/main/java/com/example/retro_care/invoice/repository/IInvoiceRepository.java
#	src/main/java/com/example/retro_care/invoice/service/IInvoiceService.java
#	src/main/java/com/example/retro_care/invoice/service/InvoiceServiceImpl.java
#	src/main/java/com/example/retro_care/kind_of_medicine/model/KindOfMedicine.java
#	src/main/java/com/example/retro_care/medicine/controller/MedicineController.java
#	src/main/java/com/example/retro_care/medicine/repository/IMedicineRepository.java
#	src/main/java/com/example/retro_care/medicine/service/IMedicineService.java
#	src/main/java/com/example/retro_care/medicine/service/Impl/MedicineService.java
#	src/main/java/com/example/retro_care/prescription/controller/PrescriptionController.java
#	src/main/java/com/example/retro_care/prescription/model/Prescription.java
#	src/main/java/com/example/retro_care/prescription/repository/IPrescriptionRepository.java
#	src/main/java/com/example/retro_care/prescription/service/PrescriptionService.java
#	src/main/java/com/example/retro_care/user/common/ValidateAppUser.java
#	src/main/java/com/example/retro_care/user/controller/AppUserController.java
#	src/main/java/com/example/retro_care/user/dto/JwtResponseUserDetails.java
#	src/main/java/com/example/retro_care/user/repository/IAppUserRepository.java
#	src/main/java/com/example/retro_care/user/service/AppUserService.java
#	src/main/java/com/example/retro_care/user/service/IAppUserService.java

+++ src/main/java/com/example/retro_care/employee/service/EmployeeService.java
@@ -24,7 +24,6 @@ public class EmployeeService implements IEmployeeService{
     @Override
     public String getNextCode() {
         String code = employeeRepository.getLastCodeEmployee();
-        System.out.println(code);
         if(code==null){
             return "NV001";
         }
@@ -36,18 +35,16 @@ public class EmployeeService implements IEmployeeService{
     /**
      * author: TanNV
      * this function use to give employee to repository and save this employee
-     *
      * @param employee
-     * @param userId
      * @return this employee have id
      */
     @Override
-    public void addEmployee(Employee employee, Long userId) {
+    public void addEmployee(Employee employee) {
 
         employeeRepository.addEmployee(employee.getCodeEmployee(),employee.getNameEmployee(),
                 employee.getAddress(),employee.getPhoneNumber(),employee.getStartDay(),
                 employee.getBirthday(),employee.getIdCard(),employee.getImage(),
-                employee.getNote(),employee.isFlagDelete(),userId);
+                employee.getNote(),employee.isFlagDelete(),employee.getAppUser().getId());
     }
   /**
      * Create: SonTT
@@ -66,12 +63,13 @@ public class EmployeeService implements IEmployeeService{
      * Date create: 15/09/2023
      * Function: search employee with id and name and with pagination
      * @param pageable
+     * @param id
      * @param name
      * @return Page
      */
     @Override
-    public Page<Employee> searchEmployee(Pageable pageable, String name) {
-        return employeeRepository.searchEmployeeByNameAndRole(pageable,name);
+    public Page<Employee> searchEmployee(Pageable pageable, Long id, String name) {
+        return employeeRepository.searchEmployeeByNameAndRole(pageable,id,name);
     }
 
     /**
@@ -81,13 +79,8 @@ public class EmployeeService implements IEmployeeService{
      * @return List
      */
     @Override
-    public Employee findEmployee(Long id) {
-        try{
-            return employeeRepository.findEmployeeById(id);
-        }catch (Exception e){
-            return null;
-        }
-
+    public List<AppRole> getRole() {
+        return employeeRepository.getRole();
     }
 
     /**
@@ -99,11 +92,11 @@ public class EmployeeService implements IEmployeeService{
      */
     @Override
     public boolean deleteEmployee(Long id) {
-        try{
+        if (employeeRepository.findEmployeeById(id) == null){
+            return false;
+        }else {
             employeeRepository.deleteEmployeeById(id);
             return true;
-        }catch (Exception e){
-            return false;
         }
     }
 
@@ -130,6 +123,6 @@ public class EmployeeService implements IEmployeeService{
         employeeRepository.updateEmployee(employee.getNameEmployee(),
                 employee.getAddress(),employee.getPhoneNumber(),employee.getStartDay(),
                 employee.getBirthday(),employee.getIdCard(),employee.getImage(),
-                employee.getNote(),employee.getId());
+                employee.getNote());
     }
 }
